refactor: complete remaining 12 code review items
Phase 1 — Config fixes:
- Fix TypeScript version mismatch in card-reader (^5.3.3 → ^5.6.0)
- Remove npm install fallback in Dockerfile.web (use npm ci only)
- Add explicit docker network to docker-compose.prod.yml
- Enable @typescript-eslint/no-explicit-any as "warn" in both apps
- Add Firefox and WebKit browser projects to Playwright config

Phase 2 — New utilities:
- Add frontend env validation (apps/web/src/lib/env.ts)
- Extract useLiffInit hook from 5 duplicate LIFF pages
- Add THAI_NAME_PREFIXES and RELATIONSHIP_OPTIONS to constants.ts
- Create backend pagination utility (parsePagination, paginatedResult)

Phase 3 — Backend refactoring:
- Split contracts.service.ts (1,196→523 lines) into 4 services:
  ContractsService (CRUD), ContractWorkflowService, ContractPaymentService,
  ContractDocumentService
- Add Prisma soft-delete middleware to auto-filter deletedAt: null
- Remove ~100 manual deletedAt: null checks across 12+ services

Phase 4 — Wire up utilities:
- Replace inline LIFF init in 5 pages with useLiffInit hook
- Replace duplicate Thai constants in 4 files with shared imports

Phase 5 — CI/CD:
- Add E2E test step to CI pipeline (Playwright + Chromium)
- Add npm audit security scanning step
- Upload Playwright report as CI artifact
