+++ .github/workflows/deploy.yml
@@ -41,6 +41,10 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
+      - name: Security audit
+        run: npm audit --audit-level=high
+        continue-on-error: true
+
       - name: Generate Prisma Client
         run: npx prisma generate --schema=apps/api/prisma/schema.prisma
 
@@ -66,6 +70,51 @@ jobs:
 
       - name: Build Web
         run: npm run build --workspace=apps/web
+
+      - name: Install Playwright Browsers
+        run: npx playwright install --with-deps chromium
+        working-directory: apps/web
+
+      - name: Seed test database
+        run: npx prisma db seed --schema=apps/api/prisma/schema.prisma
+        continue-on-error: true
+        env:
+          DATABASE_URL: postgresql://test:test@localhost:5432/test_db?schema=public
+
+      - name: Start API server
+        run: node apps/api/dist/main.js &
+        env:
+          DATABASE_URL: postgresql://test:test@localhost:5432/test_db?schema=public
+          JWT_SECRET: test-secret
+          JWT_REFRESH_SECRET: test-refresh-secret
+          JWT_EXPIRATION: 15m
+          JWT_REFRESH_EXPIRATION: 7d
+          PORT: 3000
+          NODE_ENV: test
+
+      - name: Wait for API
+        run: npx wait-on http://localhost:3000/api/health --timeout 30000
+
+      - name: Serve Web build
+        run: npx serve -s apps/web/dist -l 5173 &
+
+      - name: Wait for Web
+        run: npx wait-on http://localhost:5173 --timeout 15000
+
+      - name: Run E2E tests
+        run: npx playwright test --project=chromium
+        working-directory: apps/web
+        env:
+          BASE_URL: http://localhost:5173
+          CI: true
+
+      - name: Upload Playwright report
+        uses: actions/upload-artifact@v4
+        if: always()
+        with:
+          name: playwright-report
+          path: apps/web/playwright-report/
+          retention-days: 7
 
   build-and-push:
     name: Build & Push Docker Image
