+++ Dockerfile.web
@@ -6,7 +6,7 @@ WORKDIR /app
 COPY package.json package-lock.json* ./
 COPY apps/web/package.json apps/web/
 COPY packages/ packages/
-RUN npm ci 2>/dev/null || npm install
+RUN npm ci
 
 # ============================================
 # Stage 2: Build frontend
