+++ apps/api/.eslintrc.json
@@ -16,7 +16,7 @@
   },
   "ignorePatterns": ["dist/", "node_modules/"],
   "rules": {
-    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-explicit-any": "warn",
     "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }]
   }
 }
