+++ apps/api/src/modules/contracts/contracts.service.ts
@@ -1,11 +1,10 @@
 import { Injectable, Logger, NotFoundException, BadRequestException, ForbiddenException, InternalServerErrorException } from '@nestjs/common';
-import { PaymentMethod, PlanType, Prisma } from '@prisma/client';
+import { PlanType, Prisma } from '@prisma/client';
 import { PrismaService } from '../../prisma/prisma.service';
-import { NotificationsService } from '../notifications/notifications.service';
 import { CreateContractDto, UpdateContractDto } from './dto/contract.dto';
 import { calculateInstallment, generatePaymentSchedule } from '../../utils/installment.util';
-import { loadInstallmentConfig, resolveInstallmentParams, BUSINESS_RULES } from '../../utils/config.util';
-import { generateContractNumber, generateSaleNumber } from '../../utils/sequence.util';
+import { loadInstallmentConfig, resolveInstallmentParams } from '../../utils/config.util';
+import { generateContractNumber } from '../../utils/sequence.util';
 import {
   validateIMEI,
   validateThaiPhone,
@@ -15,14 +14,12 @@ import {
   checkRequiredDocuments,
   checkRequiredSignatures,
 } from '../../utils/validation.util';
-import * as crypto from 'crypto';
 
 @Injectable()
 export class ContractsService {
   private readonly logger = new Logger(ContractsService.name);
   constructor(
     private prisma: PrismaService,
-    private notificationsService: NotificationsService,
   ) {}
 
   async findAll(filters: {
@@ -35,7 +32,7 @@ export class ContractsService {
     limit?: number;
     salespersonId?: string;
   }) {
-    const where: Record<string, unknown> = { deletedAt: null };
+    const where: Record<string, unknown> = {};
     if (filters.status) where.status = filters.status;
     if (filters.workflowStatus) where.workflowStatus = filters.workflowStatus;
     if (filters.branchId) where.branchId = filters.branchId;
@@ -494,309 +491,6 @@ export class ContractsService {
     return this.findOne(id);
   }
 
-  // === WORKFLOW: ส่งตรวจสอบ (ตรวจ Validation ทั้งหมดก่อนส่ง) ===
-  async submitForReview(id: string, userId: string) {
-    const contract = await this.findOne(id);
-
-    if (contract.workflowStatus !== 'CREATING' && contract.workflowStatus !== 'REJECTED') {
-      throw new BadRequestException('สัญญาต้องอยู่ในสถานะ กำลังสร้าง หรือ ปฏิเสธ เท่านั้น');
-    }
-
-    // Only the salesperson who created it can submit
-    if (contract.salespersonId !== userId) {
-      throw new ForbiddenException('เฉพาะพนักงานที่สร้างสัญญาเท่านั้นที่สามารถส่งตรวจสอบ');
-    }
-
-    // Enforce mandatory steps before submit
-    // Step 1: Credit check must be approved
-    if (!contract.creditCheck || contract.creditCheck.status !== 'APPROVED') {
-      throw new BadRequestException('ต้องผ่านการตรวจเครดิตก่อนส่งตรวจสอบ (ขั้นตอนที่ 1)');
-    }
-
-    // Step 2: Required fields validation
-    const customer = contract.customer;
-    const product = contract.product;
-    const missingFields = checkRequiredContractFields({
-      customerName: customer?.name,
-      customerNationalId: customer?.nationalId,
-      customerPhone: customer?.phone,
-      customerAddressIdCard: customer?.addressIdCard,
-      customerAddressCurrent: customer?.addressCurrent,
-      references: customer?.references as any[],
-      productName: product?.name,
-      productImei: product?.imeiSerial,
-      sellingPrice: Number(contract.sellingPrice),
-      downPayment: Number(contract.downPayment),
-      totalMonths: contract.totalMonths,
-      monthlyPayment: Number(contract.monthlyPayment),
-    });
-    if (missingFields.length > 0) {
-      throw new BadRequestException(`ข้อมูลสัญญาไม่ครบ: ${missingFields.join(', ')} (ขั้นตอนที่ 2)`);
-    }
-
-    // Step 3: PDPA consent
-    if (!contract.pdpaConsentId) {
-      throw new BadRequestException('ต้องได้รับความยินยอม PDPA จากลูกค้าก่อน (ขั้นตอนที่ 3)');
-    }
-
-    // Step 5: Signatures (at minimum customer + company)
-    const customerSigned = contract.signatures?.some((s: { signerType: string }) =>
-      s.signerType === 'CUSTOMER'
-    );
-    const companySigned = contract.signatures?.some((s: { signerType: string }) =>
-      s.signerType === 'COMPANY' || s.signerType === 'STAFF'
-    );
-    if (!customerSigned || !companySigned) {
-      throw new BadRequestException('ต้องลงนามครบทั้งลูกค้าและผู้ขายก่อนส่งตรวจสอบ (ขั้นตอนที่ 5)');
-    }
-
-    // Generate contract hash for integrity
-    const contractData = JSON.stringify({
-      contractNumber: contract.contractNumber,
-      customerId: contract.customerId,
-      productId: contract.productId,
-      sellingPrice: contract.sellingPrice,
-      downPayment: contract.downPayment,
-      totalMonths: contract.totalMonths,
-      monthlyPayment: contract.monthlyPayment,
-    });
-    const contractHash = crypto.createHash('sha256').update(contractData).digest('hex');
-
-    await this.prisma.contract.update({
-      where: { id },
-      data: {
-        workflowStatus: 'PENDING_REVIEW',
-        contractHash,
-        // Set legal clause flags
-        hasOwnershipClause: true,
-        hasRepossessionClause: true,
-        hasEarlyPayoffClause: true,
-        hasNoTransferClause: true,
-        hasAcknowledgement: true,
-      },
-    });
-
-    return this.findOne(id);
-  }
-
-  // === WORKFLOW: อนุมัติสัญญา (ตรวจเอกสารครบก่อนอนุมัติ) ===
-  async approveContract(id: string, userId: string, userRole: string, reviewNotes?: string) {
-    const contract = await this.findOne(id);
-
-    if (contract.workflowStatus !== 'PENDING_REVIEW') {
-      throw new BadRequestException('สัญญาต้องอยู่ในสถานะ รอตรวจสอบ');
-    }
-
-    // Prevent self-approval: salesperson cannot approve their own contract
-    // Exception: OWNER can always approve (for small business where owner is also salesperson)
-    if (contract.salespersonId === userId && userRole !== 'OWNER') {
-      throw new ForbiddenException('ไม่สามารถอนุมัติสัญญาที่ตัวเองสร้างได้');
-    }
-
-    // Check age for guardian requirement
-    let requiresGuardian = false;
-    if (contract.customer?.birthDate) {
-      const ageCheck = checkAgeEligibility(new Date(contract.customer.birthDate));
-      requiresGuardian = ageCheck.requiresGuardian;
-    }
-
-    // Step 7: Manager ตรวจสอบเอกสารครบ
-    const docCheck = checkRequiredDocuments(
-      contract.contractDocuments || [],
-      requiresGuardian,
-    );
-    if (!docCheck.complete) {
-      const missing = docCheck.checklist
-        .filter((c) => !c.present)
-        .map((c) => c.label);
-      throw new BadRequestException(`เอกสารไม่ครบ ไม่สามารถอนุมัติได้: ${missing.join(', ')}`);
-    }
-
-    // Check signatures completeness
-    const sigCheck = checkRequiredSignatures(
-      contract.signatures || [],
-      requiresGuardian,
-    );
-    if (!sigCheck.complete) {
-      const missing = sigCheck.checklist
-        .filter((c) => !c.signed)
-        .map((c) => c.label);
-      throw new BadRequestException(`ลายเซ็นไม่ครบ ไม่สามารถอนุมัติได้: ${missing.join(', ')}`);
-    }
-
-    await this.prisma.contract.update({
-      where: { id },
-      data: {
-        workflowStatus: 'APPROVED',
-        reviewedById: userId,
-        reviewedAt: new Date(),
-        reviewNotes,
-      },
-    });
-
-    return this.findOne(id);
-  }
-
-  // === WORKFLOW: ปฏิเสธสัญญา ===
-  async rejectContract(id: string, userId: string, userRole: string, reviewNotes: string) {
-    const contract = await this.findOne(id);
-
-    if (contract.workflowStatus !== 'PENDING_REVIEW') {
-      throw new BadRequestException('สัญญาต้องอยู่ในสถานะ รอตรวจสอบ');
-    }
-
-    // OWNER can always reject (for small business where owner is also salesperson)
-    if (contract.salespersonId === userId && userRole !== 'OWNER') {
-      throw new ForbiddenException('ไม่สามารถปฏิเสธสัญญาที่ตัวเองสร้างได้');
-    }
-
-    await this.prisma.contract.update({
-      where: { id },
-      data: {
-        workflowStatus: 'REJECTED',
-        reviewedById: userId,
-        reviewedAt: new Date(),
-        reviewNotes,
-      },
-    });
-
-    return this.findOne(id);
-  }
-
-  async activate(id: string) {
-    const contract = await this.findOne(id);
-
-    // Must be APPROVED workflow and DRAFT status
-    if (contract.workflowStatus !== 'APPROVED') {
-      throw new BadRequestException('สัญญาต้องได้รับการอนุมัติก่อนเปิดใช้งาน');
-    }
-    if (contract.status !== 'DRAFT') {
-      throw new BadRequestException('สัญญาต้องอยู่ในสถานะ DRAFT');
-    }
-
-    // Verify PDPA consent exists (compliance: พ.ร.บ.คุ้มครองข้อมูลส่วนบุคคล พ.ศ. 2562)
-    if (!contract.pdpaConsentId) {
-      throw new BadRequestException('ต้องได้รับความยินยอม PDPA ก่อนเปิดใช้งานสัญญา');
-    }
-
-    // Require all required signatures (customer + company + 2 witnesses)
-    const customerSigned = contract.signatures?.some((s: { signerType: string }) => s.signerType === 'CUSTOMER');
-    const companySigned = contract.signatures?.some((s: { signerType: string }) =>
-      s.signerType === 'COMPANY' || s.signerType === 'STAFF'
-    );
-    const witness1Signed = contract.signatures?.some((s: { signerType: string }) => s.signerType === 'WITNESS_1');
-    const witness2Signed = contract.signatures?.some((s: { signerType: string }) => s.signerType === 'WITNESS_2');
-
-    if (!customerSigned || !companySigned) {
-      throw new BadRequestException('ต้องลงนามครบทั้งผู้ซื้อและผู้ขายก่อนเปิดใช้งานสัญญา');
-    }
-    if (!witness1Signed || !witness2Signed) {
-      throw new BadRequestException('ต้องมีพยานลงนามครบ 2 คนก่อนเปิดใช้งานสัญญา');
-    }
-
-    // Check guardian signature if required
-    if (contract.customer?.birthDate) {
-      const ageCheck = checkAgeEligibility(new Date(contract.customer.birthDate));
-      if (ageCheck.requiresGuardian) {
-        const guardianSigned = contract.signatures?.some((s: { signerType: string }) => s.signerType === 'GUARDIAN');
-        if (!guardianSigned) {
-          throw new BadRequestException('ลูกค้าอายุต่ำกว่า 20 ปี ต้องมีผู้ปกครองลงนาม');
-        }
-      }
-    }
-
-    // Verify product is still reserved for this contract
-    const product = await this.prisma.product.findUnique({ where: { id: contract.productId } });
-    if (!product || (product.status !== 'RESERVED' && product.status !== 'IN_STOCK')) {
-      throw new BadRequestException('สินค้าไม่พร้อมสำหรับเปิดสัญญา (อาจถูกขายหรือลบไปแล้ว)');
-    }
-
-    await this.prisma.$transaction(async (tx) => {
-      // Re-check product status inside transaction to prevent race condition
-      const prod = await tx.product.findUnique({ where: { id: contract.productId } });
-      if (!prod || (prod.status !== 'RESERVED' && prod.status !== 'IN_STOCK')) {
-        throw new BadRequestException('สินค้าไม่พร้อมสำหรับเปิดสัญญา (อาจถูกขายหรือลบไปแล้ว)');
-      }
-      // Step 8: สถานะเปลี่ยนเป็น ACTIVE → เริ่มนับงวด
-      await tx.contract.update({ where: { id }, data: { status: 'ACTIVE' } });
-      await tx.product.update({ where: { id: contract.productId }, data: { status: 'SOLD_INSTALLMENT' } });
-
-      // Auto-create Sale record for this contract activation
-      const saleNumber = await generateSaleNumber(tx);
-      await tx.sale.create({
-        data: {
-          saleNumber,
-          saleType: 'INSTALLMENT',
-          customerId: contract.customerId,
-          productId: contract.productId,
-          branchId: contract.branchId,
-          salespersonId: contract.salespersonId,
-          sellingPrice: contract.sellingPrice,
-          discount: 0,
-          netAmount: contract.sellingPrice,
-          paymentMethod: 'CASH',
-          amountReceived: contract.downPayment,
-          downPaymentAmount: contract.downPayment,
-          contractId: contract.id,
-          bundleProductIds: [],
-          notes: `สร้างอัตโนมัติจากสัญญา ${contract.contractNumber}`,
-        },
-      });
-    });
-
-    // Send LINE notification to customer (non-blocking)
-    this.sendContractActivatedNotification(contract).catch(err =>
-      this.logger.warn(`Failed to send contract activation notification: ${err?.message || err}`),
-    );
-
-    return this.findOne(id);
-  }
-
-  private async sendContractActivatedNotification(contract: any) {
-    if (!this.notificationsService) {
-      this.logger.warn(
-        `NotificationsService unavailable - cannot send activation notification for contract ${contract.contractNumber || contract.id}`,
-      );
-      return;
-    }
-    const customer = contract.customer;
-    if (!customer) return;
-
-    const firstPayment = await this.prisma.payment.findFirst({
-      where: { contractId: contract.id, installmentNo: 1 },
-      select: { dueDate: true },
-    });
-    const firstDueDate = firstPayment
-      ? new Date(firstPayment.dueDate).toLocaleDateString('th-TH')
-      : 'ตามสัญญา';
-
-    const message = [
-      `สัญญาผ่อนชำระ ${contract.contractNumber} อนุมัติแล้ว`,
-      `สินค้า: ${contract.product?.brand || ''} ${contract.product?.model || ''}`,
-      `ค่างวด: ${Number(contract.monthlyPayment).toLocaleString()} ฿/เดือน`,
-      `งวดแรก: ${firstDueDate}`,
-      `ขอบคุณที่ไว้วางใจ BESTCHOICE`,
-    ].join('\n');
-
-    const lineId = customer.lineId;
-    if (lineId) {
-      await this.notificationsService.send({
-        channel: 'LINE',
-        recipient: lineId,
-        message,
-        relatedId: contract.id,
-        fallbackPhone: customer.phone || undefined,
-      });
-    } else if (customer.phone) {
-      await this.notificationsService.send({
-        channel: 'SMS',
-        recipient: customer.phone,
-        message,
-        relatedId: contract.id,
-      });
-    }
-  }
-
   // === SOFT DELETE: ลบสัญญา (เฉพาะ CREATING/REJECTED, ห้ามลบสัญญาที่ลงนามแล้ว) ===
   async softDelete(id: string, userId: string) {
     const contract = await this.findOne(id);
@@ -825,372 +519,5 @@ export class ContractsService {
     ]);
 
     return { message: 'ลบสัญญาเรียบร้อย (soft delete)' };
-  }
-
-  async getSchedule(id: string) {
-    await this.findOne(id);
-    return this.prisma.payment.findMany({
-      where: { contractId: id },
-      orderBy: { installmentNo: 'asc' },
-    });
-  }
-
-  async getEarlyPayoffQuote(id: string) {
-    const contract = await this.findOne(id);
-    if (!['ACTIVE', 'OVERDUE', 'DEFAULT'].includes(contract.status)) {
-      throw new BadRequestException('สัญญาต้องอยู่ในสถานะ ACTIVE, OVERDUE หรือ DEFAULT');
-    }
-
-    // Count fully paid AND partially paid installments
-    const paidPayments = contract.payments.filter((p) => p.status === 'PAID');
-    const partialPayments = contract.payments.filter((p) => p.status === 'PARTIALLY_PAID');
-    const fullyPaidCount = paidPayments.length;
-    const remainingMonths = contract.totalMonths - fullyPaidCount;
-
-    if (remainingMonths <= 0) {
-      throw new BadRequestException('ไม่มีงวดค้างชำระ ไม่จำเป็นต้องปิดก่อนกำหนด');
-    }
-
-    const monthlyInterest = Number(contract.interestTotal) / contract.totalMonths;
-    // Use financedAmount (includes commission + VAT) minus interest for true principal
-    const truePrincipal = Number(contract.financedAmount) - Number(contract.interestTotal);
-    const monthlyPrincipal = truePrincipal / contract.totalMonths;
-
-    const remainingPrincipal = monthlyPrincipal * remainingMonths;
-    const remainingInterest = monthlyInterest * remainingMonths;
-    const discount = remainingInterest * BUSINESS_RULES.EARLY_PAYOFF_DISCOUNT;
-
-    // Deduct amounts already partially paid
-    const partiallyPaidAmount = partialPayments.reduce(
-      (sum, p) => sum + Number(p.amountPaid || 0), 0,
-    );
-    const totalPayoff = Math.max(0, remainingPrincipal + (remainingInterest - discount) - partiallyPaidAmount);
-
-    // Add any unpaid late fees
-    const unpaidLateFees = contract.payments
-      .filter((p) => p.status !== 'PAID')
-      .reduce((sum, p) => sum + Number(p.lateFee), 0);
-
-    return {
-      remainingMonths,
-      remainingPrincipal: Math.round(remainingPrincipal),
-      remainingInterest: Math.round(remainingInterest),
-      discount: Math.round(discount),
-      partiallyPaidCredit: Math.round(partiallyPaidAmount),
-      unpaidLateFees,
-      totalPayoff: Math.round(totalPayoff + unpaidLateFees),
-    };
-  }
-
-  async earlyPayoff(id: string, userId: string, paymentMethod: string) {
-    const quote = await this.getEarlyPayoffQuote(id);
-
-    await this.prisma.$transaction(async (tx) => {
-      // Re-verify contract status inside transaction to prevent race condition
-      const freshContract = await tx.contract.findUnique({
-        where: { id },
-        select: { status: true },
-      });
-      if (!freshContract || !['ACTIVE', 'OVERDUE', 'DEFAULT'].includes(freshContract.status)) {
-        throw new BadRequestException('สถานะสัญญาไม่อนุญาตให้ปิดก่อนกำหนด');
-      }
-
-      const unpaidPayments = await tx.payment.findMany({
-        where: { contractId: id, status: { not: 'PAID' } },
-        orderBy: { installmentNo: 'asc' },
-      });
-
-      let remainingPayoff = quote.totalPayoff;
-      for (const payment of unpaidPayments) {
-        const owed = Number(payment.amountDue) + Number(payment.lateFee) - Number(payment.amountPaid);
-        const payAmount = Math.min(remainingPayoff, owed);
-        remainingPayoff -= payAmount;
-
-        await tx.payment.update({
-          where: { id: payment.id },
-          data: {
-            status: 'PAID',
-            paidDate: new Date(),
-            amountPaid: Number(payment.amountPaid) + payAmount,
-            paymentMethod: paymentMethod as PaymentMethod,
-            recordedById: userId,
-          },
-        });
-      }
-
-      await tx.contract.update({
-        where: { id },
-        data: { status: 'EARLY_PAYOFF' },
-      });
-    });
-
-    return { ...quote, status: 'EARLY_PAYOFF' };
-  }
-
-  // ─── Document Dashboard for Manager/Admin ──────────────
-  async getDocumentDashboard(branchId?: string) {
-    const where: Record<string, unknown> = { deletedAt: null };
-    if (branchId) where.branchId = branchId;
-
-    // Get all active contracts with their documents and signatures
-    const contracts = await this.prisma.contract.findMany({
-      where,
-      include: {
-        branch: { select: { id: true, name: true } },
-        customer: { select: { name: true } },
-        eDocuments: { select: { id: true, documentType: true } },
-        signatures: { select: { id: true, signerType: true } },
-      },
-      orderBy: { createdAt: 'desc' },
-    });
-
-    const REQUIRED_DOCS = ['ID_CARD_COPY', 'KYC_SELFIE', 'DEVICE_PHOTO', 'DEVICE_IMEI_PHOTO', 'DOWN_PAYMENT_RECEIPT', 'PDPA_CONSENT'];
-    const REQUIRED_SIGS = ['CUSTOMER', 'COMPANY', 'WITNESS_1', 'WITNESS_2'];
-
-    let fullyDocumented = 0;
-    let pendingDocuments = 0;
-    let pendingSignatures = 0;
-    let pendingApproval = 0;
-    let overdueContracts = 0;
-
-    const branchMap = new Map<string, { branchId: string; branchName: string; total: number; documented: number; pendingDocs: number; pendingSigs: number }>();
-
-    for (const c of contracts) {
-      const docTypes = new Set(c.eDocuments.map((d) => d.documentType));
-      const sigTypes = new Set(c.signatures.map((s) => s.signerType as string));
-      const hasAllDocs = REQUIRED_DOCS.every((d) => docTypes.has(d));
-      const hasAllSigs = REQUIRED_SIGS.every((s) => sigTypes.has(s));
-
-      if (hasAllDocs && hasAllSigs) fullyDocumented++;
-      if (!hasAllDocs) pendingDocuments++;
-      if (!hasAllSigs) pendingSignatures++;
-      if (c.workflowStatus === 'PENDING_REVIEW' || c.workflowStatus === 'CREATING') pendingApproval++;
-      if (c.status === 'OVERDUE' || c.status === 'DEFAULT') overdueContracts++;
-
-      // By branch
-      const bId = c.branchId || 'unknown';
-      const bName = c.branch?.name || 'ไม่ระบุ';
-      if (!branchMap.has(bId)) {
-        branchMap.set(bId, { branchId: bId, branchName: bName, total: 0, documented: 0, pendingDocs: 0, pendingSigs: 0 });
-      }
-      const b = branchMap.get(bId)!;
-      b.total++;
-      if (hasAllDocs && hasAllSigs) b.documented++;
-      if (!hasAllDocs) b.pendingDocs++;
-      if (!hasAllSigs) b.pendingSigs++;
-    }
-
-    // SLA alerts: contracts waiting for approval > 24h
-    const slaAlerts = contracts
-      .filter((c) => ['PENDING_REVIEW', 'CREATING'].includes(c.workflowStatus || ''))
-      .map((c) => {
-        const hoursWaiting = Math.round((Date.now() - new Date(c.updatedAt).getTime()) / (1000 * 60 * 60));
-        return {
-          id: c.id,
-          contractNumber: c.contractNumber,
-          customerName: c.customer?.name || '',
-          workflowStatus: c.workflowStatus,
-          hoursWaiting,
-          branchName: c.branch?.name || '',
-        };
-      })
-      .filter((a) => a.hoursWaiting >= 24)
-      .sort((a, b) => b.hoursWaiting - a.hoursWaiting)
-      .slice(0, 20);
-
-    // Recent document activity from audit log (batch fetch to avoid N+1)
-    let recentActivity: Array<{ id: string; contractNumber: string; customerName: string; action: string; createdAt: string; branchName: string }> = [];
-    try {
-      const audits = await this.prisma.documentAuditLog.findMany({
-        orderBy: { createdAt: 'desc' },
-        take: 20,
-      });
-
-      // Batch-fetch all referenced contracts in one query
-      const contractIds = [...new Set(audits.map((a) => a.contractId))];
-      const auditContracts = contractIds.length > 0
-        ? await this.prisma.contract.findMany({
-            where: { id: { in: contractIds } },
-            select: {
-              id: true,
-              contractNumber: true,
-              customer: { select: { name: true } },
-              branch: { select: { name: true } },
-            },
-          })
-        : [];
-      const contractMap = new Map(auditContracts.map((c) => [c.id, c]));
-
-      recentActivity = audits.map((a) => {
-        const c = contractMap.get(a.contractId);
-        return {
-          id: a.id,
-          contractNumber: c?.contractNumber || '',
-          customerName: c?.customer?.name || '',
-          action: a.action,
-          createdAt: a.createdAt.toISOString(),
-          branchName: c?.branch?.name || '',
-        };
-      });
-    } catch {
-      // DocumentAuditLog might not exist yet
-    }
-
-    return {
-      totalContracts: contracts.length,
-      fullyDocumented,
-      pendingDocuments,
-      pendingSignatures,
-      pendingApproval,
-      overdueContracts,
-      byBranch: Array.from(branchMap.values()),
-      recentActivity,
-      slaAlerts,
-    };
-  }
-
-  // ─── QR Code Verification ────────────────────────────
-  async verifyContract(id: string, hash?: string) {
-    const contract = await this.prisma.contract.findUnique({
-      where: { id },
-      select: {
-        id: true,
-        contractNumber: true,
-        contractHash: true,
-        status: true,
-        workflowStatus: true,
-        createdAt: true,
-        totalMonths: true,
-        monthlyPayment: true,
-        customer: { select: { name: true } },
-        branch: { select: { name: true } },
-        signatures: {
-          select: { signerType: true, signerName: true, signedAt: true },
-        },
-      },
-    });
-
-    if (!contract || !contract.contractHash) {
-      return { verified: false, reason: 'ไม่พบสัญญาหรือสัญญายังไม่ได้ยืนยัน' };
-    }
-
-    const isHashValid = hash ? contract.contractHash === hash : true;
-    const signerSummary = contract.signatures.map((s) => ({
-      type: s.signerType,
-      name: s.signerName,
-      signedAt: s.signedAt,
-    }));
-
-    return {
-      verified: isHashValid,
-      reason: isHashValid ? 'สัญญาได้รับการยืนยันแล้ว' : 'Hash ไม่ตรงกัน สัญญาอาจถูกแก้ไข',
-      contract: {
-        contractNumber: contract.contractNumber,
-        status: contract.status,
-        workflowStatus: contract.workflowStatus,
-        customerName: contract.customer?.name || '',
-        branchName: contract.branch?.name || '',
-        createdAt: contract.createdAt,
-        totalMonths: contract.totalMonths,
-        monthlyPayment: contract.monthlyPayment,
-      },
-      signatures: signerSummary,
-      hash: contract.contractHash,
-    };
-  }
-
-  async getQrData(id: string) {
-    const contract = await this.prisma.contract.findUnique({
-      where: { id },
-      select: { id: true, contractNumber: true, contractHash: true },
-    });
-
-    if (!contract) throw new NotFoundException('ไม่พบสัญญา');
-
-    // QR code content: verification URL with contract ID and hash
-    const verifyUrl = `/api/contracts/${contract.id}/verify?hash=${contract.contractHash || ''}`;
-
-    return {
-      contractId: contract.id,
-      contractNumber: contract.contractNumber,
-      contractHash: contract.contractHash,
-      verifyUrl,
-      qrContent: JSON.stringify({
-        type: 'BESTCHOICE_CONTRACT',
-        id: contract.id,
-        number: contract.contractNumber,
-        hash: contract.contractHash,
-        verifyUrl,
-      }),
-    };
-  }
-
-  /** Record PDPA consent and link to contract */
-  async recordPdpaConsent(
-    contractId: string,
-    signatureImage: string,
-    req: { ip?: string; userAgent?: string },
-  ) {
-    // Validate signature image size (max 2MB base64 ≈ ~2.7MB string)
-    const MAX_SIGNATURE_SIZE = 3 * 1024 * 1024; // 3MB string length
-    if (!signatureImage) {
-      throw new BadRequestException('ต้องมีลายเซ็น');
-    }
-    if (signatureImage.length > MAX_SIGNATURE_SIZE) {
-      throw new BadRequestException('ลายเซ็นมีขนาดใหญ่เกินไป (สูงสุด 2MB)');
-    }
-
-    const contract = await this.prisma.contract.findUnique({
-      where: { id: contractId },
-      include: { customer: true },
-    });
-    if (!contract || contract.deletedAt) throw new NotFoundException('ไม่พบสัญญา');
-    if (contract.pdpaConsentId) throw new BadRequestException('สัญญานี้มี PDPA consent แล้ว');
-
-    // Get privacy notice version
-    const versionConfig = await this.prisma.systemConfig.findUnique({
-      where: { key: 'pdpa_privacy_notice_version' },
-    });
-
-    const consent = await this.prisma.pDPAConsent.create({
-      data: {
-        customerId: contract.customerId,
-        consentVersion: versionConfig?.value || '1.0',
-        privacyNoticeText: 'ยินยอมตาม พ.ร.บ.คุ้มครองข้อมูลส่วนบุคคล พ.ศ. 2562',
-        purposes: [
-          'สัญญาผ่อนชำระสินค้า',
-          'ติดตามหนี้และบริหารสัญญา',
-          'จัดทำเอกสารทางกฎหมาย',
-          'ติดต่อสื่อสารเกี่ยวกับสัญญา',
-        ],
-        status: 'GRANTED',
-        grantedAt: new Date(),
-        ipAddress: req.ip || null,
-        deviceInfo: req.userAgent || null,
-        signatureImage: signatureImage || null,
-      },
-    });
-
-    // Link consent to contract
-    await this.prisma.contract.update({
-      where: { id: contractId },
-      data: { pdpaConsentId: consent.id },
-    });
-
-    return consent;
-  }
-
-  /** Get PDPA consent for contract */
-  async getPdpaConsent(contractId: string) {
-    const contract = await this.prisma.contract.findUnique({
-      where: { id: contractId },
-      select: { pdpaConsentId: true },
-    });
-    if (!contract) throw new NotFoundException('ไม่พบสัญญา');
-    if (!contract.pdpaConsentId) return null;
-
-    return this.prisma.pDPAConsent.findUnique({
-      where: { id: contract.pdpaConsentId },
-    });
   }
 }
