+++ apps/api/src/modules/purchase-orders/purchase-orders.service.ts
@@ -772,7 +772,7 @@ export class PurchaseOrdersService {
 
     return this.prisma.$transaction(async (tx) => {
       const products = await tx.product.findMany({
-        where: { id: { in: productIds }, deletedAt: null },
+        where: { id: { in: productIds } },
       });
 
       // Validate all products are QC_PENDING
@@ -820,7 +820,7 @@ export class PurchaseOrdersService {
    * Get products pending QC (QC_PENDING status)
    */
   async getQCPending(filters: { branchId?: string; page?: number; limit?: number }) {
-    const where: Record<string, unknown> = { status: 'QC_PENDING', deletedAt: null };
+    const where: Record<string, unknown> = { status: 'QC_PENDING' };
     if (filters.branchId) where.branchId = filters.branchId;
 
     const page = Math.max(1, filters.page || 1);
