+++ apps/api/src/modules/reorder-points/reorder-points.service.ts
@@ -37,7 +37,6 @@ export class ReorderPointsService {
             category: rp.category,
             branchId: rp.branchId,
             status: 'IN_STOCK',
-            deletedAt: null,
           },
         });
         return { ...rp, currentStock, isLow: currentStock <= rp.minQuantity };
@@ -128,7 +127,6 @@ export class ReorderPointsService {
           category: rp.category,
           branchId: rp.branchId,
           status: 'IN_STOCK',
-          deletedAt: null,
         },
       });
 
@@ -303,7 +301,6 @@ export class ReorderPointsService {
           category: rp.category,
           branchId: rp.branchId,
           status: 'IN_STOCK',
-          deletedAt: null,
         },
       });
