+++ apps/api/src/modules/stock-count/stock-count.service.ts
@@ -75,7 +75,6 @@ export class StockCountService {
       const expectedProducts = await tx.product.findMany({
         where: {
           branchId: dto.branchId,
-          deletedAt: null,
           status: { in: ['IN_STOCK', 'RESERVED', 'QC_PENDING'] },
         },
         select: { id: true, status: true },
