+++ apps/card-reader/package.json
@@ -22,6 +22,6 @@
     "@types/node": "^20.11.0",
     "esbuild": "^0.24.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.6.0"
   }
 }
