+++ apps/web/.eslintrc.json
@@ -19,7 +19,7 @@
   },
   "ignorePatterns": ["dist/", "node_modules/"],
   "rules": {
-    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-explicit-any": "warn",
     "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }],
     "no-empty": "off"
   }
