+++ apps/web/playwright.config.ts
@@ -33,6 +33,14 @@ export default defineConfig({
         },
       },
     },
+    {
+      name: 'firefox',
+      use: { ...devices['Desktop Firefox'] },
+    },
+    {
+      name: 'webkit',
+      use: { ...devices['Desktop Safari'] },
+    },
   ],
   // Start dev server before running tests (if not already running)
   webServer: {
