+++ apps/web/src/main.tsx
@@ -1,3 +1,4 @@
+import '@/lib/env';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
